backend/requirements.txt
langchain_community
langchain_openai
langchain
langchain_core
langchain_experimental
langchainhub
python-dotenv
faiss-cpu
pypdf==4.3.0
pypdf2
langchain-ai21

streamlit==1.34.0

pymilvus==2.2.0
sentence-transformers
fastapi
uvicorn[standard]